feat: add cancel and save buttons to top of page
